refactor: relocate email samples and regexp test out of lib into the user module
