+++ src/module/user/domain/user/value-object/email/email.samples.ts
